Remove class-sink and **kwargs from "add()" valid arguments
There are several reasons for this restriction:
- It's more user-friendly if a argument is mistyped, as it will directly
raise an exception when "add()" is called rather that printing the error
to stderr (if "catch=True") at run-time once ".log()" is called.
- It's actually not the job of "add()" to instantiate sinks nor manage
its possible extra arguments. If the user need to do so, he can simply
wrap it with a lambda or use a "functools.partial()" function which is
purposely designed to do that.
- It simplifies both the API and the code base.
- It optmizes sink calls as most of the time "kwargs" would be empty.